+++ tests/testthat/test-create-plots-from-excel.R
@@ -22,11 +22,19 @@ simulatedScenarios <- runScenarios(
   scenarios = scenarios
 )
 
+# For compatibility with projects created with esqlabsR <5.0.1, use old data set
+# naming pattern.
+importerConfiguration <- ospsuite::loadDataImporterConfiguration(
+  configurationFilePath = projectConfiguration$dataImporterConfigurationFile
+)
+importerConfiguration$namingPattern <- "{Molecule}_{Study Id}_{Subject Id}_{Species}_{Organ}_{Compartment}_{Dose}_{Route}_{Group Id}"
+
 # Load observed data
 dataSheets <- "Laskin 1982.Group A"
 observedData <- esqlabsR::loadObservedData(
   projectConfiguration = projectConfiguration,
-  sheets = dataSheets
+  sheets = dataSheets,
+  importerConfiguration = importerConfiguration
 )
 
 dataCombinedDf <- data.frame(list(
@@ -53,8 +61,8 @@ plotConfigurationDf <- data.frame(list(
   "yUnit" = NA,
   "xAxisScale" = NA,
   "yAxisScale" = NA,
-  "xAxisLimits" = NA,
-  "yAxisLimits" = NA,
+  "xValuesLimits" = NA,
+  "yValuesLimits" = NA,
   "quantiles" = NA,
   "foldDistance" = NA
 ))
@@ -283,7 +291,7 @@ test_that("It trows an error if defined scenario is missing and stopIfNotFound i
         observedData = observedData,
         projectConfiguration = projectConfigurationLocal,
         stopIfNotFound = TRUE
-      ), regexp = messages$stopInvalidScenarioName("foo"))
+      ), regexp = messages$warningInvalidScenarioName("foo"))
     }
   )
 })